another bug fix from int rounding

+++ color_helper.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 #include "color_helper.h"
 
-struct rgb_color *make_rgb_color(int red, int green, int blue)
+struct rgb_color *make_rgb_color(float red, float green, float blue)
 {
     struct rgb_color *made_struct = (struct rgb_color *)malloc(sizeof(struct rgb_color));
     made_struct->r = red;
